Refactor: sign_up 폴더 생성 및 upload_image.dart 이름 user_verify.dart 로 변경
