+++ client/front/lib/check_image.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
-import 'upload_image.dart';
+import 'user_verify.dart';
 import 'profile.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
