+++ client/front/lib/sign_up.dart
@@ -2,7 +2,7 @@ import 'dart:async'; //Timer이용 위함.
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'custom_dialog.dart';
-import 'upload_image.dart'; // 파일 호출
+import 'user_verify.dart'; // 파일 호출
 import 'package:http/http.dart' as http;
 import 'dart:convert';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
